ADF Test

Conduct an ADF test and automatically find the first difference if the variable is non-stationary.

ADF(my_var, df, type = "type2", time_var)

Arguments

my_var
The variable to be tested, with the null hypothesis stationary (random walk) and the alternative as non-stationary.

df
The data frame where the variable is contained, e.g. my_dataframe.

type
The three types are "no constant, no trend", "constant, no trend" and "constant, trend". The default is "type2", but there is also "type1" and "type3".

time_var
If the supplied data frame has previously been converted to time series using make_ts and/or the "zoo" package, then a time variable must be specified. The variable in the dataset which marks the year of the observations. E.g. "year"

Value

The result of the ADF test, and the first difference of the variable if it is stationary

Examples

file = system.file("extdata", "hseinv.txt", package="Kenometrics")
data <- read.delim(file)
data <- ADF("linvpc", data)

#> Augmented Dickey-Fuller Test 
#> alternative: stationary 
#>  
#> Type 1: no drift no trend 
#>      lag    ADF p.value
#> [1,]   0 -1.280   0.216
#> [2,]   1 -1.062   0.294
#> [3,]   2 -1.134   0.268
#> [4,]   3 -0.804   0.386
#> Type 2: with drift no trend 
#>      lag   ADF p.value
#> [1,]   0 -3.00  0.0465
#> [2,]   1 -3.33  0.0226
#> [3,]   2 -3.25  0.0266
#> [4,]   3 -1.54  0.5020
#> Type 3: with drift and trend 
#>      lag   ADF p.value
#> [1,]   0 -3.67  0.0394
#> [2,]   1 -4.59  0.0100
#> [3,]   2 -4.83  0.0100
#> [4,]   3 -3.33  0.0807
#> ---- 
#> Note: in fact, p.value = 0.01 means p.value <= 0.01 
#> [1] "beta 1  is not significant"
#> [1] "beta 2  is not significant"
#> [1] "beta 3  is not significant"
#> [1] "beta 4  is significant"
#> [1] "No change has been made to the data frame"

data <- ADF("lpop", data, type="type1")

#> Augmented Dickey-Fuller Test 
#> alternative: stationary 
#>  
#> Type 1: no drift no trend 
#>      lag    ADF p.value
#> [1,]   0 23.499   0.990
#> [2,]   1  0.955   0.905
#> [3,]   2  0.384   0.748
#> [4,]   3  0.351   0.739
#> Type 2: with drift no trend 
#>      lag    ADF p.value
#> [1,]   0 -12.64  0.0100
#> [2,]   1  -2.64  0.0963
#> [3,]   2  -1.66  0.4580
#> [4,]   3  -1.80  0.4032
#> Type 3: with drift and trend 
#>      lag   ADF p.value
#> [1,]   0 -1.98   0.570
#> [2,]   1 -1.95   0.583
#> [3,]   2 -2.66   0.308
#> [4,]   3 -3.06   0.154
#> ---- 
#> Note: in fact, p.value = 0.01 means p.value <= 0.01 
#> [1] "beta 1  is significant"
#> [1] "lpop is not stationary"
#> [1] "beta 2  is significant"
#> [1] "beta 3  is significant"
#> [1] "beta 4  is significant"
#> [1] "The 1st difference has been appended to the data frame"