Chow Test

Conduct a chow test on a desired model for a structural break in either the intercept, slope, or both.

chow(possible_break, time_var = "year", df, model, type)

Arguments

possible_break
The year (point) in which the structural break is believed to have occured, e.g. 1969.

time_var
The variable which marks the data/time of the observations.

df
The data frame which holds the variables. e.g. cement_data

model
The model which the chow test is to be conducted on. Enter this in string form, i.e. "lpop ~ price, L(linvpc,0:2) + p"

type
The type of structural break to be assessed; a break in the break the "intercept", "slope", or "both".

Value

The output from print, giving the result of the chow test.

Examples

file = system.file("extdata", "hseinv.txt", package="Kenometrics")
data <- read.delim(file)
chow(1960, "year", data, "linvpc ~ lpop", "both")

#> linvpc ~ lpop + D + Dx
#> <environment: 0x00000000283f7c30>
#> 
#> Time series regression with "numeric" data:
#> Start = 1, End = 42
#> 
#> Call:
#> dynlm::dynlm(formula = new_model, data = df)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.46825 -0.06333 -0.00423  0.08076  0.28312 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)
#> (Intercept) -6.450623   7.424541  -0.869    0.390
#> lpop         0.473731   0.619654   0.765    0.449
#> D           -8.519504  14.453440  -0.589    0.559
#> Dx           0.004326   0.007403   0.584    0.562
#> 
#> Residual standard error: 0.1457 on 38 degrees of freedom
#> Multiple R-squared:  0.339,	Adjusted R-squared:  0.2868 
#> F-statistic: 6.495 on 3 and 38 DF,  p-value: 0.001173
#> 
#> [1] "The dummies are NOT jointly significant; a structural break has NOT occured"

#chow(1960, "Time", my_data, "lpop ~ price, L(linvpc,0:2) + p", "intercept")